Extract _settings.py from _legacy.py
